PP-10063: Remove code relating to old registration journey
Deleted routes, controllers and services.
