+++ .secrets.baseline
@@ -733,7 +733,7 @@
         "filename": "test/integration/login.controller.ft.test.js",
         "hashed_secret": "0a6f4e1251c72c635e2bdfc6e1aa46d336d80d4f",
         "is_verified": false,
-        "line_number": 237
+        "line_number": 235
       }
     ],
     "test/integration/register-user.controller.ft.test.js": [
@@ -902,5 +902,5 @@
       }
     ]
   },
-  "generated_at": "2022-12-06T17:16:05Z"
+  "generated_at": "2022-12-08T12:33:28Z"
 }
