--- app/controllers/register-user.controller.test.js
@@ -1,204 +1,0 @@
-'use strict'
-
-const sinon = require('sinon')
-const proxyquire = require('proxyquire')
-
-const userFixtures = require('../../test/fixtures/user.fixtures')
-const inviteFixtures = require('../../test/fixtures/invite.fixtures')
-const paths = require('../paths')
-const User = require('../models/User.class')
-const { RegistrationSessionMissingError, RESTClientError, ExpiredInviteError } = require('../errors')
-const { expect } = require('chai')
-
-describe('Register user controller', () => {
-  const email = 'invited-user@example.com'
-  const inviteCode = 'a-code'
-  const serviceExternalId = 'a-service-id'
-  const otpCode = '123 456'
-  const expectedSanitisedCode = '123456'
-
-  const verifyOtpSuccessStub = sinon.spy(() => Promise.resolve())
-  const completeInviteSuccessStub = sinon.spy(() => Promise.resolve(inviteFixtures.validInviteCompleteResponse({
-    service_external_id: serviceExternalId
-  })))
-  const flashSpy = sinon.spy()
-
-  let req, res, next
-
-  beforeEach(() => {
-    req = {
-      register_invite: { code: inviteCode, email },
-      user: new User(userFixtures.validUserResponse({ email })),
-      body: {
-        'verify-code': otpCode
-      },
-      flash: flashSpy
-    }
-    res = {
-      redirect: sinon.spy(),
-      render: sinon.spy(),
-      setHeader: sinon.spy(),
-      status: sinon.spy()
-    }
-    next = sinon.spy()
-    verifyOtpSuccessStub.resetHistory()
-    completeInviteSuccessStub.resetHistory()
-    flashSpy.resetHistory()
-  })
-
-  describe('Subscribe service', () => {
-    function getController (mockCompleteInvite) {
-      return proxyquire('./register-user.controller', {
-        '../services/user-registration.service': {
-          completeInvite: mockCompleteInvite
-        }
-      })
-    }
-
-    describe('Valid invite for user', () => {
-      it('should accept invite and redirect to "My services', async () => {
-        const controller = getController(completeInviteSuccessStub)
-        await controller.subscribeService(req, res, next)
-        sinon.assert.calledWith(completeInviteSuccessStub, inviteCode, 'SMS')
-        sinon.assert.calledWith(flashSpy, 'inviteSuccessServiceId', serviceExternalId)
-        sinon.assert.calledWith(completeInviteSuccessStub, inviteCode)
-      })
-    })
-
-    describe('The email on the user and invite are the same but with different case', () => {
-      it('should accept invite and redirect to "My services', async () => {
-        req.register_invite.email = 'Invited-User@example.com'
-
-        const controller = getController(completeInviteSuccessStub)
-        await controller.subscribeService(req, res, next)
-        sinon.assert.called(completeInviteSuccessStub)
-        sinon.assert.calledWith(flashSpy, 'inviteSuccessServiceId', serviceExternalId)
-        sinon.assert.calledWith(completeInviteSuccessStub, inviteCode)
-      })
-    })
-
-    describe('Logged in user is not the invited user', () => {
-      it('should redirect to "My services" without accepting invite', async () => {
-        req.register_invite.email = 'a-different-user@example.com'
-
-        const controller = getController(completeInviteSuccessStub)
-        await controller.subscribeService(req, res, next)
-        sinon.assert.notCalled(completeInviteSuccessStub)
-        sinon.assert.calledWith(res.redirect, 303, paths.serviceSwitcher.index)
-      })
-    })
-
-    describe('Cookie details are missing', () => {
-      it('should call next with error', async () => {
-        delete req.register_invite
-
-        const controller = getController(completeInviteSuccessStub)
-        await controller.subscribeService(req, res, next)
-        sinon.assert.notCalled(completeInviteSuccessStub)
-        sinon.assert.calledWith(next, sinon.match.instanceOf(RegistrationSessionMissingError))
-      })
-    })
-
-    describe('Invitation is expired', () => {
-      it('should render error page', async () => {
-        const completeInviteStub = sinon.stub().throws({ errorCode: 410 })
-        const controller = getController(completeInviteStub)
-        await controller.subscribeService(req, res, next)
-        sinon.assert.calledWith(res.status, 410)
-        sinon.assert.calledWithMatch(res.render, 'error', { message: 'This invitation is no longer valid' })
-      })
-    })
-  })
-
-  describe('Submit otp verify', () => {
-    function getController (verifyOtpSpy, completeInviteSpy) {
-      return proxyquire('./register-user.controller', {
-        '../services/user-registration.service.js': {
-          verifyOtp: verifyOtpSpy,
-          completeInvite: completeInviteSpy
-        }
-      })
-    }
-
-    describe('Submitted OTP code is correct', () => {
-      it('should redirect to logUserIn with a 303', async () => {
-        const controllerWithVerifyOtpSuccess = getController(verifyOtpSuccessStub, completeInviteSuccessStub)
-        await controllerWithVerifyOtpSuccess.submitOtpVerify(req, res, next)
-
-        sinon.assert.calledWith(verifyOtpSuccessStub, inviteCode, expectedSanitisedCode)
-        sinon.assert.calledWith(completeInviteSuccessStub, inviteCode)
-        sinon.assert.calledWith(res.redirect, 303, paths.registerUser.logUserIn)
-      })
-    })
-
-    describe('Call to verifyOtp returns a 401 response', () => {
-      it('should set an error message in the cookie and redirect to registerUser.otpVerify with a 303', async () => {
-        const error = new RESTClientError('An error', 'adminusers', 401)
-        const verifyOtpFailureStub = () => Promise.reject(error)
-
-        const controllerWithVerifyOtpError = getController(verifyOtpFailureStub, completeInviteSuccessStub)
-        await controllerWithVerifyOtpError.submitOtpVerify(req, res, next)
-
-        expect(req.register_invite).to.have.property('recovered')
-        expect(req.register_invite.recovered).to.deep.equal({
-          errors: {
-            securityCode: 'The security code you’ve used is incorrect or has expired'
-          }
-        })
-        sinon.assert.calledWith(res.redirect, 303, paths.registerUser.otpVerify)
-      })
-    })
-
-    describe('Call to verifyOtp returns a 410 response', () => {
-      it('should call next with with an ExpiredInviteError', async () => {
-        const error = new RESTClientError('An error', 'adminusers', 410)
-        const verifyOtpFailureStub = () => Promise.reject(error)
-
-        const controllerWithVerifyOtpError = getController(verifyOtpFailureStub, completeInviteSuccessStub)
-        await controllerWithVerifyOtpError.submitOtpVerify(req, res, next)
-
-        const expectedError = sinon.match.instanceOf(ExpiredInviteError)
-          .and(sinon.match.has('message', `Invite with code ${inviteCode} has expired`))
-        sinon.assert.calledWith(next, expectedError)
-      })
-    })
-
-    describe('Call to verifyOtp returns some other status code', () => {
-      it('should call next with the error', async () => {
-        const error = new RESTClientError('An error', 'adminusers', 500)
-        const verifyOtpFailureStub = () => Promise.reject(error)
-
-        const controllerWithVerifyOtpError = getController(verifyOtpFailureStub, completeInviteSuccessStub)
-        await controllerWithVerifyOtpError.submitOtpVerify(req, res, next)
-
-        sinon.assert.calledWith(next, error)
-      })
-    })
-
-    describe('Call to completeInvite returns a 410 response', () => {
-      it('should call next with with an ExpiredInviteError', async () => {
-        const error = new RESTClientError('An error', 'adminusers', 410)
-        const completeInviteFailureStub = () => Promise.reject(error)
-
-        const controllerWithCompleteOtpError = getController(verifyOtpSuccessStub, completeInviteFailureStub)
-        await controllerWithCompleteOtpError.submitOtpVerify(req, res, next)
-
-        const expectedError = sinon.match.instanceOf(ExpiredInviteError)
-          .and(sinon.match.has('message', `Invite with code ${inviteCode} has expired`))
-        sinon.assert.calledWith(next, expectedError)
-      })
-    })
-
-    describe('Call to completeInvite returns some other status code', () => {
-      it('should call next with the error', async () => {
-        const error = new RESTClientError('An error', 'adminusers', 500)
-        const completeInviteFailureStub = () => Promise.reject(error)
-
-        const controllerWithCompleteOtpError = getController(verifyOtpSuccessStub, completeInviteFailureStub)
-        await controllerWithCompleteOtpError.submitOtpVerify(req, res, next)
-
-        sinon.assert.calledWith(next, error)
-      })
-    })
-  })
-})
