+++ app/routes.test.js
@@ -19,15 +19,7 @@ const pathsNotRequiringAuthentication = [
   paths.user.passwordRequested,
   paths.user.forgottenPasswordReset,
   paths.inviteValidation.validateInvite,
-  paths.registerUser.registration,
   paths.registerUser.subscribeService,
-  paths.registerUser.otpVerify,
-  paths.registerUser.reVerifyPhone,
-  paths.registerUser.logUserIn,
-  paths.selfCreateService.register,
-  paths.selfCreateService.confirm,
-  paths.selfCreateService.otpVerify,
-  paths.selfCreateService.otpResend,
   paths.healthcheck.path,
   paths.staticPaths.naxsiError,
   paths.register.email,
