--- app/views/self-create-service/confirm.njk
@@ -1,13 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-
-{% block pageTitle %}
-  Account confirmation - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-<div id="display-email-sent" class="govuk-grid-column-two-thirds main-content">
-    <h1 class="govuk-heading-l page-title">Check your email</h1>
-    <p class="govuk-body-l">An email has been sent to {{requesterEmail}}.</p>
-    <p class="govuk-body">Click the link in the email to continue your registration.</p>
-</div>
-{% endblock %}
