--- app/views/self-create-service/register.njk
@@ -1,83 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-{% from "../macro/error-summary.njk" import errorSummary %}
-
-{% block pageTitle %}
-  Create an account - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-<div class="govuk-grid-column-two-thirds">
-  {{ errorSummary ({
-    errors: errors,
-    hrefs: {
-      email: '#email',
-      telephoneNumber: '#telephone-number',
-      password: '#password'
-    }
-  }) }}
-  <h1 class="govuk-heading-l page-title">Create an account</h1>
-  <form action="{{routes.selfCreateService.register}}" method="post" id="submit-service-creation" class="form submit-registration">
-    <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}"/>
-    <div class="govuk-form-group {% if errors['email'] %} govuk-form-group--error {% endif %}">
-        <label class="govuk-label govuk-label--s" for="email">
-            Email
-        </label>
-        <div id="email-hint" class="govuk-hint">
-            <p class="govuk-hint">Must be from a government organisation</p>
-        </div>
-        {% if errors['email'] %}
-            <span id="email-error" class="govuk-error-message">
-                <span class="govuk-visually-hidden">Error:</span> {{errors["email"]}}. If you cannot create an account using your public sector email address, please <a href="https://www.payments.service.gov.uk/support/" id='support-link'>contact support</a>.
-            </span>
-        {% endif %}
-        <input class="govuk-input govuk-input--width-20 {% if errors['email'] %} govuk-input--error {% endif %}"
-            id="email" name="email" type="email" aria-describedby="email-hint"
-            value="{{email}}"/>
-    </div>
-    {{
-      govukInput({
-        id: "telephone-number",
-        name: "telephone-number",
-        type: "tel",
-        autocomplete: "tel",
-        classes: "govuk-input--width-20",
-        label: {
-            text: "Mobile number",
-            classes: "govuk-label--s"
-        },
-        value: telephoneNumber,
-        hint: {
-          text: "We’ll send you a security code by text message"
-        },
-        errorMessage: { text: errors.telephoneNumber } if errors.telephoneNumber else false
-      })
-    }}
-    {{
-      govukInput({
-        id: "password",
-        name: "password",
-        type: "password",
-        autocomplete: "new-password",
-        classes: "govuk-input--width-20",
-        label: {
-            text: "Password",
-            classes: "govuk-label--s"
-        },
-        hint: {
-          text: "Must be at least 10 characters"
-        },
-        errorMessage: { text: errors.password } if errors.password else false
-      })
-    }}
-    {{
-      govukButton({
-        text: "Continue",
-        classes: "button",
-        attributes: {
-          id: "continue"
-        }
-      })
-    }}
-  </form>
-</div>
-{% endblock %}
