--- app/views/self-create-service/resend-otp.njk
@@ -1,47 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-{% from "../macro/error-summary.njk" import errorSummary %}
-
-{% block pageTitle %}
-  Resend security code - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-<div id="display_otp_resend" class="govuk-grid-column-two-thirds">
-  {{ errorSummary ({
-    errors: errors,
-    hrefs: {
-      telephoneNumber: '#telephone-number'
-    }
-  }) }}
-
-  <h1 class="form-title govuk-heading-l">Check your mobile number</h1>
-  <p class="govuk-body">Check your mobile phone number is correct, then resend the security code.</p>
-  <form action="{{routes.selfCreateService.otpResend}}" class="form" method="post" id="otp-resend-form">
-    <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}" />
-    {{
-      govukInput({
-        id: "telephone-number",
-        name: "telephone-number",
-        type: "tel",
-        autocomplete: "tel",
-        classes: "govuk-input--width-20",
-        label: {
-            text: "Mobile number",
-            classes: "govuk-label--s"
-        },
-        value: telephoneNumber,
-        errorMessage: { text: errors.telephoneNumber } if errors.telephoneNumber else false
-      })
-    }}
-    {{
-      govukButton({
-        text: "Resend security code",
-        classes: "button",
-        attributes: {
-          id: "resend-code"
-        }
-      })
-    }}
-  </form>
-</div>
-{% endblock %}
