--- app/views/self-create-service/verify-otp.njk
@@ -1,49 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-{% from "../macro/error-summary.njk" import errorSummary %}
-
-{% block pageTitle %}
-  Enter security code - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-<div id="display_otp_verify" class="govuk-grid-column-two-thirds">
-  {{ errorSummary ({
-    errors: errors,
-    hrefs: {
-      securityCode: '#verify-code'
-    }
-  }) }}
-
-  <h1 class="form-title govuk-heading-l">Check your phone</h1>
-  <p class="govuk-body">We’ve sent you a text message with a security code</p>
-  <form action="{{routes.selfCreateService.otpVerify}}" class="form" method="post" class="submit-two-fa" id="verify-phone-form">
-    <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}" />
-    {{
-      govukInput({
-        id: "verify-code",
-        name: "verify-code",
-        classes: "govuk-input--width-10",
-        label: {
-            text: "Security code",
-            classes: "govuk-label--s"
-        },
-        attributes: {
-          autocomplete: "off",
-          autofocus: true
-        },
-        errorMessage: { text: errors.securityCode } if errors.securityCode else false
-      })
-    }}
-    {{
-      govukButton({
-        text: "Continue",
-        classes: 'button',
-        attributes: {
-          id: "continue"
-        }
-      })
-    }}
-  </form>
-  <p class="govuk-body"><a href="{{routes.selfCreateService.otpResend}}" class="resend-otp-link govuk-link govuk-link govuk-link--no-visited-state">Not received a text message?</a></p>
-</div>
-{% endblock %}
