--- app/views/user-registration/re-verify-phone.njk
@@ -1,52 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-{% from "../macro/error-summary.njk" import errorSummary %}
-
-{% block pageTitle %}
-  Resend security code - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-  <div class="govuk-grid-column-two-thirds">
-    {{ errorSummary ({
-      errors: errors,
-      hrefs: {
-        telephoneNumber: '#telephone-number'
-      }
-    }) }}
-
-    <h1 class="form-title govuk-heading-l">Resend security code</h1>
-    <p class="govuk-body">Text messages sometimes take a few minutes to arrive. If you did not receive the text, you can resend it.</p>
-    <p class="govuk-body">If you no longer have access to the phone with the number registered for this service, speak to your service manager to reset the number.</p>
-    <form action="{{routes.registerUser.reVerifyPhone}}" class="form" method="post" id="otp-send-again">
-      <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}" />
-      <input type="hidden" name="send" value="true"/>
-      {{
-        govukInput({
-          id: "telephone-number",
-          name: "telephone-number",
-          type: "tel",
-          autocomplete: "tel",
-          classes: "govuk-input--width-20",
-          label: {
-              text: "Mobile number",
-              classes: "govuk-label--s"
-          },
-          value: telephone_number,
-          hint: {
-            text: "We’ll send you a security code by text message"
-          },
-          errorMessage: { text: errors.telephoneNumber } if errors.telephoneNumber else false
-        })
-      }}
-      {{
-        govukButton({
-          text: "Resend security code",
-          classes: "button",
-          attributes: {
-            id: "resend-code"
-          }
-        })
-      }}
-    </form>
-  </div>
-{% endblock %}
