--- app/views/user-registration/register.njk
@@ -1,67 +1,0 @@
-{% extends "../layout-logged-out.njk" %}
-{% from "../macro/error-summary.njk" import errorSummary %}
-
-{% block pageTitle %}
-  Create an account - GOV.UK Pay
-{% endblock %}
-
-{% block mainContent %}
-  <div class="govuk-grid-column-two-thirds">
-    {{ errorSummary ({
-    errors: errors,
-    hrefs: {
-      telephoneNumber: '#telephone-number',
-      password: '#password'
-    }
-  }) }}
-    <h1 class="govuk-heading-l page-title">Create an account</h1>
-    <p class="govuk-body-l" id="email-display">Your account will be created with this email: {{email}}</p>
-    <form action="{{routes.registerUser.registration}}" method="post" id="submit-registration" class="form submit-registration">
-      <input id="csrf" name="csrfToken" type="hidden" value="{{csrf}}"/>
-      {{
-        govukInput({
-          id: "telephone-number",
-          name: "telephone-number",
-          type: "tel",
-          autocomplete: "tel",
-          classes: "govuk-input--width-20",
-          label: {
-              text: "Mobile number",
-              classes: "govuk-label--s"
-          },
-          value: telephone_number,
-          hint: {
-            text: "We’ll send you a security code by text message"
-          },
-          errorMessage: { text: errors.telephoneNumber } if errors.telephoneNumber else false
-        })
-      }}
-      {{
-        govukInput({
-          id: "password",
-          name: "password",
-          type: "password",
-          autocomplete: "new-password",
-          classes: "govuk-input--width-20",
-          label: {
-              text: "Password",
-              classes: "govuk-label--s"
-          },
-          hint: {
-            text: "Must be at least 10 characters"
-          },
-          errorMessage: { text: errors.password } if errors.password else false
-        })
-      }}
-      {{
-        govukButton({
-          text: "Continue",
-          classes: "button",
-          attributes: {
-            id: "continue"
-          }
-        })
-      }}
-    </form>
-  </div>
-{% endblock %}
